Reddit Ovarra/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.10.0
requests==2.32.3
google-generativeai==0.8.3
textblob==0.18.0
